+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] ; 2025-07-03
+
+### Changed
+
+- Timeseries splitting logic to be robust to the presence of other index columns
+- Added `fit_and_predict()` method to adapter class
+
 ## [0.4.5] ; 2025-06-27
 
 ### Changed